Crate anvil_zksync_types

Modules

api

numbers

traces

Structs

L2TxBuilder

TransactionPriority

Metric value for the priority of a transaction.

Enums

LogLevel

Log filter level for the node.

Numeric

Helper type to be able to parse both integers (as u64) and hex strings (as U64) depending on the user input.

ShowGasDetails

ShowStorageLogs

ShowVMDetails

TransactionOrder

Modes that determine the transaction ordering of the mempool